Implement extract reviews and metadata jobs

+++ src/main/scala/org/amazon/reviews/types/Review.scala
@@ -5,7 +5,7 @@ import java.sql.Timestamp
 case class Review(reviewerID: String,
                   asin: String,
                   reviewerName: String,
-                  helpful: Seq[Int],
+                  helpful: Seq[Long],
                   reviewText: String,
                   overall: Double,
                   summary: String,
